backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
cssselect2==0.7.0
distro==1.9.0
Django==5.1.1
django-extensions==3.2.3
django-cors-headers==4.4.0
docopt==0.6.2
fonttools==4.55.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.1
markdown2==2.5.1
md2pdf==1.0.1
multidict==6.1.0
openai==1.54.4
packaging==24.2
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.15.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
yarl==1.17.1
zopfli==0.2.3.post1
google-cloud-storage==3.1.0
